# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Affinity Risk Solutions for AAA Clubs - Telematics Insurance Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /facilities:
    get:
      operationId: listFacilities
      tags: [facilities]
      summary: List towing facilities, optionally filtered by club
      parameters:
        - name: clubId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of facilities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Facility"
  /facilities/{facilityId}:
    get:
      operationId: getFacility
      tags: [facilities]
      summary: Get facility by ID
      parameters:
        - name: facilityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Facility
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Facility"
  /facilities/{facilityId}/summary:
    get:
      operationId: getFacilitySummary
      tags: [facilities]
      summary: Get aggregated risk summary for a facility
      parameters:
        - name: facilityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Facility summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacilitySummary"

  # ─── Drivers ─────────────────────────────────────────────────────────────────
...
  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: List vehicles, optionally filtered by facility
      parameters:
        - name: facilityId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
  /vehicles/{vehicleId}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      summary: Get vehicle by ID
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{vehicleId}/telematics:
    get:
      operationId: getVehicleTelematics
      tags: [vehicles]
      summary: Get recent telematics events for a vehicle
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Telematics events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TelematicsEvent"

  # ─── Risk Scores ──────────────────────────────────────────────────────────────
  /risk/facility/{facilityId}:
    get:
      operationId: getFacilityRiskScore
      tags: [risk]
      summary: Get current risk score for a facility
      parameters:
        - name: facilityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Risk score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskScore"
  /risk/driver/{driverId}:
    get:
      operationId: getDriverRiskScore
      tags: [risk]
      summary: Get current risk score for a driver
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Risk score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskScore"
  /risk/vehicle/{vehicleId}:
    get:
      operationId: getVehicleRiskScore
      tags: [risk]
      summary: Get current risk score for a vehicle
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Risk score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskScore"
  /risk/suggestions/{driverId}:
    get:
      operationId: getDriverSuggestions
      tags: [risk]
      summary: Get behavior modification suggestions for a driver
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BehaviorSuggestion"

  # ─── Certificates of Insurance ────────────────────────────────────────────────
...
components:
  schemas:
...
    Facility:
      type: object
      required: [id, clubId, name, address, city, state, ownerName, riskScore, riskTier, status, totalDrivers, totalVehicles]
      properties:
        id:
          type: integer
        clubId:
          type: integer
        clubName:
          type: string
        name:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        ownerName:
          type: string
        ownerEmail:
          type: ["string", "null"]
        ownerPhone:
          type: ["string", "null"]
        riskScore:
          type: number
        riskTier:
          type: string
          enum: [low, moderate, high, critical]
        status:
          type: string
          enum: [active, suspended, pending]
        totalDrivers:
          type: integer
        totalVehicles:
          type: integer
        certStatus:
          type: string
          enum: [current, expiring_soon, expired, missing]
        createdAt:
          type: string
    FacilitySummary:
      type: object
      required: [facilityId, avgRiskScore, riskTier, totalDrivers, totalVehicles, activeVehicles, totalMiles, hardBrakingEvents, harshAccelerationEvents, speedingEvents, openAccidents, premiumEstimate]
      properties:
        facilityId:
          type: integer
        avgRiskScore:
          type: number
        riskTier:
          type: string
        totalDrivers:
          type: integer
        totalVehicles:
          type: integer
        activeVehicles:
          type: integer
        totalMiles:
          type: number
        hardBrakingEvents:
          type: integer
        harshAccelerationEvents:
          type: integer
        speedingEvents:
          type: integer
        openAccidents:
          type: integer
        premiumEstimate:
          type: number
...
    Vehicle:
      type: object
      required: [id, facilityId, make, model, year, licensePlate, vin, type, status, riskScore, riskTier, totalMiles, assignedDriverId]
      properties:
        id:
          type: integer
        facilityId:
          type: integer
        facilityName:
          type: ["string", "null"]
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        licensePlate:
          type: string
        vin:
          type: string
        type:
          type: string
          enum: [flatbed, wheel_lift, integrated, rotator, heavy_duty]
        status:
          type: string
          enum: [active, inactive, maintenance]
        riskScore:
          type: number
        riskTier:
          type: string
          enum: [low, moderate, high, critical]
        totalMiles:
          type: number
        assignedDriverId:
          type: ["integer", "null"]
        assignedDriverName:
          type: ["string", "null"]
        telematicsDeviceId:
          type: ["string", "null"]
        createdAt:
          type: string
    TelematicsEvent:
      type: object
      required: [id, vehicleId, driverId, eventType, severity, latitude, longitude, speed, timestamp]
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        driverId:
          type: ["integer", "null"]
        eventType:
          type: string
          enum: [hard_braking, harsh_acceleration, speeding, phone_usage, sharp_turn, fatigue_detected, accident]
        severity:
          type: string
          enum: [low, medium, high, critical]
        latitude:
          type: number
        longitude:
          type: number
        speed:
          type: number
        timestamp:
          type: string
        notes:
          type: ["string", "null"]
    RiskScore:
      type: object
      required: [entityId, entityType, score, tier, change30d, factors]
      properties:
        entityId:
          type: integer
        entityType:
          type: string
          enum: [facility, driver, vehicle]
        score:
          type: number
        tier:
          type: string
          enum: [low, moderate, high, critical]
        change30d:
          type: number
        factors:
          type: array
          items:
            $ref: "#/components/schemas/RiskFactor"
        calculatedAt:
          type: string
    RiskFactor:
      type: object
      required: [name, impact, value]
      properties:
        name:
          type: string
        impact:
          type: string
          enum: [low, medium, high]
        value:
          type: string
        description:
          type: ["string", "null"]
    BehaviorSuggestion:
      type: object
      required: [id, driverId, category, priority, title, description, expectedImpact]
      properties:
        id:
          type: integer
        driverId:
          type: integer
        category:
          type: string
          enum: [speeding, braking, acceleration, phone_usage, fatigue, route_planning]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        title:
          type: string
        description:
          type: string
        expectedImpact:
          type: string
...