# lib/api-zod/src/generated/types/riskScore.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Affinity Risk Solutions for AAA Clubs - Telematics Insurance Platform
 * OpenAPI spec version: 0.1.0
 */
import type { RiskFactor } from './riskFactor';
import type { RiskScoreEntityType } from './riskScoreEntityType';
import type { RiskScoreTier } from './riskScoreTier';

export interface RiskScore {
  entityId: number;
  entityType: RiskScoreEntityType;
  score: number;
  tier: RiskScoreTier;
  change30d: number;
  factors: RiskFactor[];
  calculatedAt?: string;
}