# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Affinity Risk Solutions for AAA Clubs - Telematics Insurance Platform
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleRiskTier } from './vehicleRiskTier';
import type { VehicleStatus } from './vehicleStatus';
import type { VehicleType } from './vehicleType';

export interface Vehicle {
  id: number;
  facilityId: number;
  /** @nullable */
  facilityName?: string | null;
  make: string;
  model: string;
  year: number;
  licensePlate: string;
  vin: string;
  type: VehicleType;
  status: VehicleStatus;
  riskScore: number;
  riskTier: VehicleRiskTier;
  totalMiles: number;
  /** @nullable */
  assignedDriverId: number | null;
  /** @nullable */
  assignedDriverName?: string | null;
  /** @nullable */
  telematicsDeviceId?: string | null;
  createdAt?: string;
}